feat: use float32

+++ unstructured/partition/pdf_image/pdfminer_processing.py
@@ -114,7 +114,7 @@ def _create_text_region(x1, y1, x2, y2, coef, text, source, region_class):
 def get_coords_from_bboxes(bboxes) -> np.ndarray:
     """convert a list of boxes's coords into np array"""
     # preallocate memory
-    coords = np.zeros((len(bboxes), 4), dtype=np.float16)
+    coords = np.zeros((len(bboxes), 4), dtype=np.float32)
 
     for i, bbox in enumerate(bboxes):
         coords[i, :] = [bbox.x1, bbox.y1, bbox.x2, bbox.y2]
